stock_prediction_project/requirements.txt
pyspark==3.4.1
PyYAML==6.0
numpy==1.24.4

kafka-python==2.0.2 
elasticsearch==8.11.0 
requests==2.31.0